+++ appGUI/PlotCanvas.py
@@ -58,7 +58,8 @@ class PlotCanvas(QtCore.QObject, VisPyCanvas):
 
         if theme == 'white':
             self.line_color = (0.3, 0.0, 0.0, 1.0)
-            self.rect_hud_color = Color('#0000FF10')
+            # self.rect_hud_color = Color('#0000FF10')
+            self.rect_hud_color = Color('#80808040')
             self.text_hud_color = 'black'
         else:
             self.line_color = (0.4, 0.4, 0.4, 1.0)