+++ CHANGELOG.md
@@ -11,6 +11,8 @@ CHANGELOG for FlatCAM beta
 
 - fixed the reconnection of the mouse events when canceling setting the origin
 - disabled the selection of a Graphic Engine until matplotlib will have support for PyQt6 framework
+- removed patch to fix scaling on HIDPI displays since now the fix is implemented in VisPy v0.8.1
+- modified a patch for VisPy Infinite Line Visual that set the line width, in order to make it compliant with newer versions of OpenGL
 
 29.08.2021